+++ jOOQ-test/src/org/jooq/test/jOOQIngresTest.java
@@ -537,7 +537,7 @@ public class jOOQIngresTest extends jOOQAbstractTest<
     }
 
     @Override
-    protected TableField<XUnusedRecord, Byte> TDirectory_IS_DIRECTORY() {
+    protected TableField<XUnusedRecord, Integer> TDirectory_IS_DIRECTORY() {
         return null;
     }
